Moves existing CLI tests into own suite
